Backend/FoodNStuff/requirements.txt
Django
psycopg2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
django-cors-headers
django-filter
